Fixed lint and relocated worker
